Extract getUnitTestBreakdown into shared lib/testTypes
The pass/fail/error counting was copy-pasted in both the agent Tests tab
and the /tests Runs table. Move it to lib/testTypes (already the shared
home for tests-list helpers), taking a minimal result shape, so the two
views stay in sync.
